Delete lost scratch orgs
Add `--ignorepool` flag to `org:scratchorg:delete` which allows deletion of
scratch orgs created by a specified signup email, without deleting pooled
scratch orgs. Useful for deleting lost scratch orgs, which are scratch orgs that
have been created, but not committed to a pool.

+++ src/commands/sfpowerkit/org/scratchorg/delete.ts
@@ -33,6 +33,15 @@ export default class Delete extends SfdxCommand {
       exclusive: ["email"],
       description: messages.getMessage("usernameFlagDescription"),
     }),
+    ignorepool: flags.boolean({
+      required: false,
+      dependsOn: ["email"],
+      description: messages.getMessage("ignorePoolFlagDescription")
+    }),
+    dryrun: flags.boolean({
+      required: false,
+      description: messages.getMessage("dryRunFlagDescription")
+    })
   };
 
   public async run(): Promise<AnyJson> {
@@ -63,14 +72,17 @@ export default class Delete extends SfdxCommand {
       );
       this.ux.table(info.records, [
         "Id",
+        "ScratchOrg",
         "SignupUsername",
         "SignupEmail",
         "ExpirationDate",
       ]);
 
-      let scratchOrgIds: string[] = info.records.map((elem) => elem.Id);
-      await ScratchOrgUtils.deleteScratchOrg(this.hubOrg, scratchOrgIds);
-      this.ux.log("Scratch Org(s) deleted successfully.");
+      if (!this.flags.dryrun) {
+        let scratchOrgIds: string[] = info.records.map((elem) => elem.Id);
+        await ScratchOrgUtils.deleteScratchOrg(this.hubOrg, scratchOrgIds);
+        this.ux.log("Scratch Org(s) deleted successfully.");
+      }
     } else {
       this.ux.log(
         `No Scratch Org(s) found for the given ${
@@ -89,7 +101,7 @@ export default class Delete extends SfdxCommand {
     email: string,
     username: string
   ): Promise<any> {
-    let query = `SELECT Id, SignupUsername, SignupEmail, ExpirationDate FROM ActiveScratchOrg`;
+    let query = `SELECT Id, ScratchOrg, SignupUsername, SignupEmail, ExpirationDate FROM ActiveScratchOrg`;
 
     if (username) {
       query = `${query} WHERE SignupUsername = '${username}'`;
@@ -97,8 +109,19 @@ export default class Delete extends SfdxCommand {
       query = `${query} WHERE SignupEmail = '${email}'`;
     }
 
+    if (this.flags.ignorepool && !username) {
+      const orgIds = await this.getOrgIdOfPooledScratchOrgs();
+      const collection = orgIds.map((id) => `'${id}'`).toString();
+      query += ` AND ScratchOrg NOT IN (${collection})`;
+    }
+    console.log("query:", query);
     const scratch_orgs = (await conn.query(query)) as any;
 
     return scratch_orgs;
+  }
+
+  private async getOrgIdOfPooledScratchOrgs(): Promise<string[]> {
+    const results = await ScratchOrgUtils.getScratchOrgsByTag(null, this.hubOrg, false, false);
+    return results.records.map((record) => record.ScratchOrg)
   }
 }
